+++ src/main/java/com/attackRange/attackRange_config.java
@@ -1,11 +1,11 @@
-package com.example;
+package com.attackRange;
 
 import net.runelite.client.config.Config;
 import net.runelite.client.config.ConfigGroup;
 import net.runelite.client.config.ConfigItem;
 
 @ConfigGroup("example")
-public interface ExampleConfig extends Config
+public interface attackRange_config extends Config
 {
 	@ConfigItem(
 		keyName = "greeting",
