+++ src/test/java/com/attackRange/attackRange_plugin_test.java
@@ -1,13 +1,13 @@
-package com.example;
+package com.attackRange;
 
 import net.runelite.client.RuneLite;
 import net.runelite.client.externalplugins.ExternalPluginManager;
 
-public class ExamplePluginTest
+public class attackRange_plugin_test
 {
 	public static void main(String[] args) throws Exception
 	{
-		ExternalPluginManager.loadBuiltin(ExamplePlugin.class);
+		ExternalPluginManager.loadBuiltin(attackRange_plugin.class);
 		RuneLite.main(args);
 	}
 }
